refactor: Move DataSourceConfig into the component where it plays a role.
